requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
celery==5.3.4
redis==5.0.1
asyncpg==0.29.0
sqlalchemy==2.0.23
alembic==1.12.1
tortoise-orm==0.20.0
aerich==0.7.2
boto3==1.29.7
botocore==1.32.7
aiofiles==23.2.0
httpx==0.25.2
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1 